refactor(node): move anomaly and deciders implementations under cuvis_ai.node
Continues the deprecation begun in d49783b: real implementations now live
under cuvis_ai/node/anomaly/ and cuvis_ai/node/deciders/, while the legacy
cuvis_ai/{anomaly,deciders}/* modules become thin shims that emit a
DeprecationWarning and re-export from the new locations. Top-level
cuvis_ai.node namespace now exposes BinaryDecider, DeepSVDDProjection,
LADGlobal, QuantileBinaryDecider, RXGlobal, RXPerBatch, TwoStageBinaryDecider,
and ZScoreNormalizerGlobal directly. The plugin manifest is updated to
register classes at their new paths.
