+++ configs/plugins/cuvis_ai_builtin.yaml
@@ -4,17 +4,17 @@ plugins:
   cuvis_ai_builtin:
     path: "../.."
     provides:
-      - cuvis_ai.anomaly.deep_svdd.DeepSVDDCenterTracker
-      - cuvis_ai.anomaly.deep_svdd.DeepSVDDProjection
-      - cuvis_ai.anomaly.deep_svdd.DeepSVDDScores
-      - cuvis_ai.anomaly.deep_svdd.ZScoreNormalizerGlobal
-      - cuvis_ai.anomaly.lad_detector.LADGlobal
-      - cuvis_ai.anomaly.rx_detector.RXBase
-      - cuvis_ai.anomaly.rx_detector.RXGlobal
-      - cuvis_ai.anomaly.rx_detector.RXPerBatch
-      - cuvis_ai.deciders.binary_decider.BinaryDecider
-      - cuvis_ai.deciders.binary_decider.QuantileBinaryDecider
-      - cuvis_ai.deciders.two_stage_decider.TwoStageBinaryDecider
+      - cuvis_ai.node.anomaly.deep_svdd.DeepSVDDCenterTracker
+      - cuvis_ai.node.anomaly.deep_svdd.DeepSVDDProjection
+      - cuvis_ai.node.anomaly.deep_svdd.DeepSVDDScores
+      - cuvis_ai.node.anomaly.deep_svdd.ZScoreNormalizerGlobal
+      - cuvis_ai.node.anomaly.lad_detector.LADGlobal
+      - cuvis_ai.node.anomaly.rx_detector.RXBase
+      - cuvis_ai.node.anomaly.rx_detector.RXGlobal
+      - cuvis_ai.node.anomaly.rx_detector.RXPerBatch
+      - cuvis_ai.node.deciders.binary_decider.BinaryDecider
+      - cuvis_ai.node.deciders.binary_decider.QuantileBinaryDecider
+      - cuvis_ai.node.deciders.two_stage_decider.TwoStageBinaryDecider
       - cuvis_ai.node.anomaly_visualization.AnomalyMask
       - cuvis_ai.node.anomaly_visualization.BBoxesOverlayNode
       - cuvis_ai.node.anomaly_visualization.ChannelSelectorFalseRGBViz
