+++ cuvis_ai/node/__init__.py
@@ -1,5 +1,12 @@
 """Convenience exports for node base classes and marker mixins."""
 
+from cuvis_ai.node.anomaly import (
+    DeepSVDDProjection,
+    LADGlobal,
+    RXGlobal,
+    RXPerBatch,
+    ZScoreNormalizerGlobal,
+)
 from cuvis_ai.node.anomaly_visualization import (
     AnomalyMask,
     BBoxesOverlayNode,
@@ -34,6 +41,7 @@ from cuvis_ai.node.channel_selector import (
 from cuvis_ai.node.colormap import ScalarHSVColormapNode
 from cuvis_ai.node.compositing import ROIZoomNode
 from cuvis_ai.node.conversion import DecisionToMask
+from cuvis_ai.node.deciders import BinaryDecider, QuantileBinaryDecider, TwoStageBinaryDecider
 from cuvis_ai.node.dimensionality_reduction import PCA, TrainablePCA
 from cuvis_ai.node.json_file import (
     CocoTrackBBoxWriter,
@@ -148,4 +156,12 @@ __all__ = [
     "TrackingOverlayNode",
     "TrackingResultsReader",
     "TrainablePCA",
+    "BinaryDecider",
+    "DeepSVDDProjection",
+    "LADGlobal",
+    "QuantileBinaryDecider",
+    "RXGlobal",
+    "RXPerBatch",
+    "TwoStageBinaryDecider",
+    "ZScoreNormalizerGlobal",
 ]
